requirements.txt
setuptools
wheel
Django==3.2
djangorestframework==3.12
boto3==1.17
azure-mgmt-resource==16.1.0
PyQt5==5.15.10
psycopg2-binary